refactoring components -> folders
